+++ pay-service/src/main/java/ru/yandex/practicum/pay/PayController.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.pay;
 
 import io.swagger.v3.oas.annotations.Parameter;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -24,7 +23,7 @@ public class PayController implements PaymentApi {
     public Mono<ResponseEntity<UserBalance>> payPost(@RequestBody Mono<PayRequest> payRequest, @Parameter(hidden = true) final ServerWebExchange exchange) {
 
         return payRequest.map(request -> {
-            Integer userId = request.getUserId();
+            Long userId = request.getUserId();
 
             System.out.println("POST pay for user "+ userId);
